First attempt at fixing pdf rotation bug. See #129
A document may be generated in a specific way and then rotated in a pdf
editor. When this happens the height and width stored in the file do not
actually change, which leads to a canvas with the original height and
width, but the content, being rotated, appears squeezed within the page.

This commit attempts to avoid this side effect, but requires additional
tweaking to center the canvas and thorough testing.